fix HideLauncherAds
not the best way but it works for now. 
as side effect we still have the placeholder as leftover and can scroll to it. not so good! 
further we should make this optional since we already introduced a setting for it.

funny thing is, even though every thing works i don't get any output from XposedBridge.log if i put it in the hook?!